feat: moves buildCanonicalfunction into metaTags folder
