--- packages/osc-ecommerce/app/utils/buildCanonicalUrl.ts
@@ -1,6 +1,0 @@
-export function buildCanonicalUrl(request: Request) {
-    const url = new URL(request?.url);
-    const canonicalUrl = url?.href;
-
-    return canonicalUrl;
-}
